JENA-1430: Quad loading for in-memory assemblers
