+++ jena-arq/src/main/java/org/apache/jena/sparql/core/assembler/InMemDatasetAssembler.java
@@ -24,13 +24,10 @@ import static org.apache.jena.riot.RDFDataMgr.read;
 import static org.apache.jena.sparql.core.assembler.AssemblerUtils.setContext;
 import static org.apache.jena.sparql.core.assembler.DatasetAssemblerVocab.pGraphName;
 import static org.apache.jena.sparql.core.assembler.DatasetAssemblerVocab.pNamedGraph;
-import static org.apache.jena.sparql.util.graph.GraphUtils.getAsStringValue;
-import static org.apache.jena.sparql.util.graph.GraphUtils.multiValueAsString;
-import static org.apache.jena.sparql.util.graph.GraphUtils.multiValueResource;
+import static org.apache.jena.sparql.util.graph.GraphUtils.*;
 
 import org.apache.jena.assembler.Assembler;
 import org.apache.jena.assembler.Mode;
-import org.apache.jena.assembler.assemblers.AssemblerBase;
 import org.apache.jena.query.Dataset;
 import org.apache.jena.rdf.model.Resource;
 import org.apache.jena.system.Txn;
@@ -38,32 +35,30 @@ import org.apache.jena.system.Txn;
 /**
  * An {@link Assembler} that creates in-memory {@link Dataset}s.
  */
-public class InMemDatasetAssembler extends AssemblerBase implements Assembler {
+public class InMemDatasetAssembler extends DatasetAssembler {
 
     public static Resource getType() {
-        return DatasetAssemblerVocab.tDatasetTxnMem ;
+        return DatasetAssemblerVocab.tMemoryDataset ;
     }
     
 	@Override
 	public Dataset open(final Assembler assembler, final Resource root, final Mode mode) {
-		checkType(root, DatasetAssemblerVocab.tDatasetTxnMem);
+		checkType(root, DatasetAssemblerVocab.tMemoryDataset);
 		final Dataset dataset = createTxnMem();
 		setContext(root, dataset.getContext());
 
 		Txn.executeWrite(dataset, ()->{ 
-    		// Load data into the default graph
-		    // This also loads quads into the dataset.
-    		multiValueAsString(root, data)
-    		    .forEach(dataURI -> read(dataset, dataURI));
-    
-    		// load data into named graphs
-    		multiValueResource(root, pNamedGraph).forEach(namedGraphResource -> {
-    			final String graphName = getAsStringValue(namedGraphResource, pGraphName);
-    			if (namedGraphResource.hasProperty(data)) {
-    			    multiValueAsString(namedGraphResource, data)
-    			        .forEach(namedGraphData -> read(dataset.getNamedModel(graphName), namedGraphData));
-    			}
-    		});
+        		// Load data into the default graph or quads into the dataset.
+        		multiValueAsString(root, data)
+        		    .forEach(dataURI -> read(dataset, dataURI));
+        
+        		// load data into named graphs
+        		multiValueResource(root, pNamedGraph).forEach(namedGraphResource -> {
+        			final String graphName = getAsStringValue(namedGraphResource, pGraphName);
+        			if (namedGraphResource.hasProperty(data))
+        			    multiValueAsString(namedGraphResource, data)
+        			        .forEach(namedGraphData -> read(dataset.getNamedModel(graphName), namedGraphData));
+        		});
 		});
 		return dataset;
 	}
