+++ jena-arq/src/test/java/org/apache/jena/sparql/core/assembler/TestDatasetAssembler.java
@@ -1,0 +1,122 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.jena.sparql.core.assembler;
+
+import static java.nio.file.Files.createTempFile;
+import static org.apache.jena.assembler.JA.data;
+import static org.apache.jena.assembler.Mode.DEFAULT;
+import static org.apache.jena.rdf.model.ModelFactory.createDefaultModel;
+import static org.apache.jena.riot.Lang.NQUADS;
+import static org.apache.jena.riot.RDFDataMgr.write;
+import static org.apache.jena.vocabulary.RDF.type;
+
+import java.io.*;
+import java.nio.file.Path;
+
+import org.apache.jena.assembler.JA;
+import org.apache.jena.assembler.exceptions.CannotConstructException;
+import org.apache.jena.query.Dataset;
+import org.apache.jena.rdf.model.*;
+import org.apache.jena.sparql.core.*;
+import org.apache.jena.sparql.sse.SSE;
+import org.apache.jena.sparql.util.IsoMatcher;
+import org.apache.jena.system.Txn;
+import org.junit.Assert;
+import org.junit.Test;
+
+public class TestDatasetAssembler extends Assert {
+    
+    protected Resource assemblerType() {
+        return DatasetAssemblerVocab.tDataset;
+    };
+    
+    protected DatasetAssembler createAssembler() {
+        return new DatasetAssembler();
+    }
+    
+    Dataset assemble(final Resource example) {
+        Model model = example.getModel() ;
+        model.setNsPrefix("ja", JA.getURI()) ;
+        final DatasetAssembler testAssembler = createAssembler();
+        return testAssembler.open(testAssembler, example, DEFAULT);
+    }
+
+    @Test
+    public void emptyDataset() {
+        final Model model = createDefaultModel();
+        final Resource empty = model.createResource("test:empty");
+        empty.addProperty(type, assemblerType()) ;
+        Dataset dataset = assemble(empty) ;
+        assertFalse(dataset.asDatasetGraph().find().hasNext()) ;
+    }
+
+    @Test(expected = CannotConstructException.class)
+    public void wrongKindOfAssemblerDefinition() {
+        final Model model = createDefaultModel();
+        final Resource badExample = model.createResource("test:badExample");
+        assemble(badExample);
+    }
+
+    @Test
+    public void directDataQuadsURI() throws IOException {
+    	// first make a file of quads to load later
+    	final Model model = createDefaultModel();
+    	final Path quads = createTempFile("quadExample", ".nq");
+    	final Resource quadsURI = model.createResource(quads.toFile().toURI().toString());
+    	final Resource simpleExample = model.createResource("test:simpleExample");
+    	simpleExample.addProperty(type, assemblerType());
+    	simpleExample.addProperty(data, quadsURI);
+    
+    	final DatasetGraph dsg = DatasetGraphFactory.createTxnMem();
+        Txn.executeWrite(dsg, () ->
+            model.listStatements().mapWith(Statement::asTriple)
+                .mapWith(t -> new Quad(quadsURI.asNode(), t)).forEachRemaining(dsg::add));
+    
+    	try (OutputStream out = new FileOutputStream(quads.toFile())) {
+    		write(out, dsg, NQUADS);
+    	}
+    
+    	final Dataset dataset = assemble(simpleExample);
+    	final Model assembledDefaultModel = dataset.getDefaultModel();
+    	final Model assembledNamedModel = dataset.getNamedModel(quadsURI.getURI());
+    	assertTrue(assembledDefaultModel.isEmpty());
+    	assertTrue(assembledNamedModel.contains(assembledNamedModel.createStatement(simpleExample, data, quadsURI)));
+    }
+
+    @Test
+    public void directDataQuadsString() throws IOException {
+        // first make a file of quads to load later
+        DatasetGraph dsgData = DatasetGraphFactory.createTxnMem();
+        dsgData.add(SSE.parseQuad("(_ :s1 :p1 :o1)"));
+        dsgData.add(SSE.parseQuad("(:gn :s2 :p2 :o2)"));
+        Path quads = createTempFile("quadExample2", ".nq");
+        String dataFileName = quads.toFile().toURI().toString();
+        try (OutputStream out = new FileOutputStream(quads.toFile())) {
+            write(out, dsgData, NQUADS);
+        }
+        
+        Model assemblerModel = createDefaultModel();
+        Resource simpleExample2 = assemblerModel.createResource("test:simpleExample2");
+        simpleExample2.addProperty(type, assemblerType()) ;
+        simpleExample2.addProperty(data, dataFileName);
+        
+        final Dataset dataset = assemble(simpleExample2);
+        
+        assertTrue(IsoMatcher.isomorphic(dsgData, dataset.asDatasetGraph()));
+    }
+}
